Refactor auth integration tests into modular components
