+++ src/services_v2/auth/run_integration_tests.py
@@ -3,724 +3,64 @@
 V2 Authentication Integration Test Runner
 ========================================
 
-Main script to run comprehensive integration tests for the V2 authentication system.
-Executes all test suites and generates detailed reports.
+Orchestrates setup, execution, and reporting for the V2 authentication
+integration tests.
 """
 
-import sys
-import os
 import time
-import json
-import logging
-
-from src.utils.stability_improvements import stability_manager, safe_import
-from datetime import datetime, timedelta
+from datetime import datetime
 from pathlib import Path
+import sys
 
 # Add src to path for imports
-sys.path.append(str(Path(__file__).parent.parent.parent))
-
-try:
-    from auth_service import AuthService
-    from auth_integration_tester import AuthIntegrationTester
-    from auth_performance_monitor import AuthPerformanceMonitor
-
-    INTEGRATION_AVAILABLE = True
-except ImportError as e:
-    print(f"Warning: Integration components not available: {e}")
-    INTEGRATION_AVAILABLE = False
-
-
-def setup_logging():
-    """Setup comprehensive logging for integration testing"""
-    logging.basicConfig(
-        level=logging.INFO,
-        format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
-        handlers=[
-            logging.FileHandler("auth_integration_tests.log"),
-            logging.StreamHandler(sys.stdout),
-        ],
-    )
-    return logging.getLogger(__name__)
-
-
-def run_basic_functionality_tests(
-    auth_service: AuthService, logger: logging.Logger
-) -> dict:
-    """Run basic functionality tests"""
-    logger.info("🧪 Running Basic Functionality Tests")
-    logger.info("-" * 40)
-
-    test_results = {
-        "tests_run": 0,
-        "tests_passed": 0,
-        "tests_failed": 0,
-        "test_details": [],
-    }
-
-    # Test 1: Basic authentication
-    try:
-        start_time = time.time()
-        result = auth_service.authenticate_user_v2(
-            "admin", "secure_password_123", "127.0.0.1", "test_agent"
-        )
-        duration = time.time() - start_time
-
-        if result.status.value == "SUCCESS":
-            test_results["tests_passed"] += 1
-            logger.info(f"✅ Basic Authentication: PASS ({duration:.3f}s)")
-        else:
-            test_results["tests_failed"] += 1
-            logger.error(
-                f"❌ Basic Authentication: FAIL - Status: {result.status.value}"
-            )
-
-        test_results["tests_run"] += 1
-        test_results["test_details"].append(
-            {
-                "test": "Basic Authentication",
-                "status": "PASS" if result.status.value == "SUCCESS" else "FAIL",
-                "duration": duration,
-                "details": str(result.status.value),
-            }
-        )
-
-    except Exception as e:
-        test_results["tests_failed"] += 1
-        test_results["tests_run"] += 1
-        logger.error(f"❌ Basic Authentication: ERROR - {e}")
-        test_results["test_details"].append(
-            {
-                "test": "Basic Authentication",
-                "status": "ERROR",
-                "duration": 0,
-                "details": str(e),
-            }
-        )
-
-    # Test 2: Invalid credentials
-    try:
-        start_time = time.time()
-        result = auth_service.authenticate_user_v2(
-            "admin", "wrong_password", "127.0.0.1", "test_agent"
-        )
-        duration = time.time() - start_time
-
-        if result.status.value != "SUCCESS":
-            test_results["tests_passed"] += 1
-            logger.info(f"✅ Invalid Credentials: PASS ({duration:.3f}s)")
-        else:
-            test_results["tests_failed"] += 1
-            logger.error(f"❌ Invalid Credentials: FAIL - Should have failed")
-
-        test_results["tests_run"] += 1
-        test_results["test_details"].append(
-            {
-                "test": "Invalid Credentials",
-                "status": "PASS" if result.status.value != "SUCCESS" else "FAIL",
-                "duration": duration,
-                "details": str(result.status.value),
-            }
-        )
-
-    except Exception as e:
-        test_results["tests_failed"] += 1
-        test_results["tests_run"] += 1
-        logger.error(f"❌ Invalid Credentials: ERROR - {e}")
-        test_results["test_details"].append(
-            {
-                "test": "Invalid Credentials",
-                "status": "ERROR",
-                "duration": 0,
-                "details": str(e),
-            }
-        )
-
-    # Test 3: Performance metrics
-    try:
-        metrics = auth_service.get_performance_metrics()
-        if metrics and "total_attempts" in metrics:
-            test_results["tests_passed"] += 1
-            logger.info(
-                f"✅ Performance Metrics: PASS - {metrics['total_attempts']} attempts"
-            )
-        else:
-            test_results["tests_failed"] += 1
-            logger.error(f"❌ Performance Metrics: FAIL - No metrics available")
-
-        test_results["tests_run"] += 1
-        test_results["test_details"].append(
-            {
-                "test": "Performance Metrics",
-                "status": "PASS" if metrics and "total_attempts" in metrics else "FAIL",
-                "duration": 0,
-                "details": str(metrics) if metrics else "No metrics",
-            }
-        )
-
-    except Exception as e:
-        test_results["tests_failed"] += 1
-        test_results["tests_run"] += 1
-        logger.error(f"❌ Performance Metrics: ERROR - {e}")
-        test_results["test_details"].append(
-            {
-                "test": "Performance Metrics",
-                "status": "ERROR",
-                "duration": 0,
-                "details": str(e),
-            }
-        )
-
-    # Test 4: Security status
-    try:
-        security_status = auth_service.get_security_status()
-        if security_status and "security_level" in security_status:
-            test_results["tests_passed"] += 1
-            logger.info(
-                f"✅ Security Status: PASS - Level: {security_status['security_level']}"
-            )
-        else:
-            test_results["tests_failed"] += 1
-            logger.error(f"❌ Security Status: FAIL - No security info")
-
-        test_results["tests_run"] += 1
-        test_results["test_details"].append(
-            {
-                "test": "Security Status",
-                "status": "PASS"
-                if security_status and "security_level" in security_status
-                else "FAIL",
-                "duration": 0,
-                "details": str(security_status)
-                if security_status
-                else "No security info",
-            }
-        )
-
-    except Exception as e:
-        test_results["tests_failed"] += 1
-        test_results["tests_run"] += 1
-        logger.error(f"❌ Security Status: ERROR - {e}")
-        test_results["test_details"].append(
-            {
-                "test": "Security Status",
-                "status": "ERROR",
-                "duration": 0,
-                "details": str(e),
-            }
-        )
-
-    return test_results
-
-
-def run_performance_tests(auth_service: AuthService, logger: logging.Logger) -> dict:
-    """Run performance and stress tests"""
-    logger.info("🚀 Running Performance Tests")
-    logger.info("-" * 40)
-
-    test_results = {
-        "tests_run": 0,
-        "tests_passed": 0,
-        "tests_failed": 0,
-        "test_details": [],
-    }
-
-    # Test 1: Single authentication performance
-    try:
-        start_time = time.time()
-        result = auth_service.authenticate_user_v2(
-            "admin", "secure_password_123", "127.0.0.1", "perf_test_agent"
-        )
-        duration = time.time() - start_time
-
-        if duration < 1.0:  # Should complete within 1 second
-            test_results["tests_passed"] += 1
-            logger.info(f"✅ Single Auth Performance: PASS ({duration:.3f}s)")
-        else:
-            test_results["tests_failed"] += 1
-            logger.warning(f"⚠️ Single Auth Performance: SLOW ({duration:.3f}s)")
-
-        test_results["tests_run"] += 1
-        test_results["test_details"].append(
-            {
-                "test": "Single Auth Performance",
-                "status": "PASS" if duration < 1.0 else "FAIL",
-                "duration": duration,
-                "details": f"Duration: {duration:.3f}s",
-            }
-        )
-
-    except Exception as e:
-        test_results["tests_failed"] += 1
-        test_results["tests_run"] += 1
-        logger.error(f"❌ Single Auth Performance: ERROR - {e}")
-        test_results["test_details"].append(
-            {
-                "test": "Single Auth Performance",
-                "status": "ERROR",
-                "duration": 0,
-                "details": str(e),
-            }
-        )
-
-    # Test 2: Multiple rapid authentications
-    try:
-        start_time = time.time()
-        successful_auths = 0
-
-        for i in range(10):
-            try:
-                result = auth_service.authenticate_user_v2(
-                    f"perf_user_{i}",
-                    "secure_password_123",
-                    f"192.168.1.{i}",
-                    "perf_test_agent",
-                )
-                if result.status.value == "SUCCESS":
-                    successful_auths += 1
-            except Exception:
-                pass  # Continue with next iteration
-
-        total_time = time.time() - start_time
-        throughput = successful_auths / total_time if total_time > 0 else 0
-
-        if throughput > 1.0:  # Should handle at least 1 auth per second
-            test_results["tests_passed"] += 1
-            logger.info(f"✅ Throughput Test: PASS - {throughput:.2f} auths/sec")
-        else:
-            test_results["tests_failed"] += 1
-            logger.warning(f"⚠️ Throughput Test: FAIL - {throughput:.2f} auths/sec")
-
-        test_results["tests_run"] += 1
-        test_results["test_details"].append(
-            {
-                "test": "Throughput Test",
-                "status": "PASS" if throughput > 1.0 else "FAIL",
-                "duration": total_time,
-                "details": f"Throughput: {throughput:.2f} auths/sec, Success: {successful_auths}/10",
-            }
-        )
-
-    except Exception as e:
-        test_results["tests_failed"] += 1
-        test_results["tests_run"] += 1
-        logger.error(f"❌ Throughput Test: ERROR - {e}")
-        test_results["test_details"].append(
-            {
-                "test": "Throughput Test",
-                "status": "ERROR",
-                "duration": 0,
-                "details": str(e),
-            }
-        )
-
-    # Test 3: Error handling under load
-    try:
-        start_time = time.time()
-        error_count = 0
-
-        for i in range(20):
-            try:
-                result = auth_service.authenticate_user_v2(
-                    f"error_user_{i}",
-                    "wrong_password",
-                    f"192.168.1.{i}",
-                    "error_test_agent",
-                )
-                if result.status.value == "SYSTEM_ERROR":
-                    error_count += 1
-            except Exception:
-                error_count += 1
-
-        total_time = time.time() - start_time
-
-        # Should handle errors gracefully without crashing
-        if (
-            error_count < 20
-        ):  # Some errors are expected, but not all should be system errors
-            test_results["tests_passed"] += 1
-            logger.info(f"✅ Error Handling: PASS - {error_count}/20 system errors")
-        else:
-            test_results["tests_failed"] += 1
-            logger.warning(f"⚠️ Error Handling: FAIL - {error_count}/20 system errors")
-
-        test_results["tests_run"] += 1
-        test_results["test_details"].append(
-            {
-                "test": "Error Handling",
-                "status": "PASS" if error_count < 20 else "FAIL",
-                "duration": total_time,
-                "details": f"System errors: {error_count}/20",
-            }
-        )
-
-    except Exception as e:
-        test_results["tests_failed"] += 1
-        test_results["tests_run"] += 1
-        logger.error(f"❌ Error Handling: ERROR - {e}")
-        test_results["test_details"].append(
-            {
-                "test": "Error Handling",
-                "status": "ERROR",
-                "duration": 0,
-                "details": str(e),
-            }
-        )
-
-    return test_results
-
-
-def run_integration_tests(auth_service: AuthService, logger: logging.Logger) -> dict:
-    """Run integration tests with other systems"""
-    logger.info("🔗 Running Integration Tests")
-    logger.info("-" * 40)
-
-    test_results = {
-        "tests_run": 0,
-        "tests_passed": 0,
-        "tests_failed": 0,
-        "test_details": [],
-    }
-
-    # Test 1: Message queue integration
-    try:
-        # This would test integration with the message queue system
-        # For now, we'll test that the auth service can handle integration scenarios
-        test_results["tests_passed"] += 1
-        logger.info(f"✅ Message Queue Integration: PASS - Integration ready")
-
-        test_results["tests_run"] += 1
-        test_results["test_details"].append(
-            {
-                "test": "Message Queue Integration",
-                "status": "PASS",
-                "duration": 0,
-                "details": "Integration framework ready",
-            }
-        )
-
-    except Exception as e:
-        test_results["tests_failed"] += 1
-        test_results["tests_run"] += 1
-        logger.error(f"❌ Message Queue Integration: ERROR - {e}")
-        test_results["test_details"].append(
-            {
-                "test": "Message Queue Integration",
-                "status": "ERROR",
-                "duration": 0,
-                "details": str(e),
-            }
-        )
-
-    # Test 2: Agent coordinator integration
-    try:
-        # This would test integration with the agent coordinator
-        # For now, we'll test that the auth service can handle coordination scenarios
-        test_results["tests_passed"] += 1
-        logger.info(f"✅ Agent Coordinator Integration: PASS - Integration ready")
-
-        test_results["tests_run"] += 1
-        test_results["test_details"].append(
-            {
-                "test": "Agent Coordinator Integration",
-                "status": "PASS",
-                "duration": 0,
-                "details": "Integration framework ready",
-            }
-        )
-
-    except Exception as e:
-        test_results["tests_failed"] += 1
-        test_results["tests_run"] += 1
-        logger.error(f"❌ Agent Coordinator Integration: ERROR - {e}")
-        test_results["test_details"].append(
-            {
-                "test": "Agent Coordinator Integration",
-                "status": "ERROR",
-                "duration": 0,
-                "details": str(e),
-            }
-        )
-
-    return test_results
-
-
-def run_comprehensive_integration_suite(logger: logging.Logger) -> dict:
-    """Run the comprehensive integration test suite"""
-    logger.info("🚀 Running Comprehensive Integration Test Suite")
-    logger.info("=" * 60)
-
-    if not INTEGRATION_AVAILABLE:
-        logger.warning(
-            "⚠️ Integration components not available, running basic tests only"
-        )
-        return run_basic_tests_only(logger)
-
-    try:
-        # Initialize integration tester
-        integration_tester = AuthIntegrationTester()
-
-        # Run tests via new orchestrator
-        start_time = time.time()
-        report = integration_tester.run()
-        total_time = time.time() - start_time
-
-        test_results = {
-            "tests_run": report.summary["total"],
-            "tests_passed": report.summary["passed"],
-            "tests_failed": report.summary["failed"],
-            "tests_error": 0,
-            "total_time": total_time,
-            "integration_status": {},
-            "performance_metrics": {},
-            "test_details": [
-                {
-                    "test": r.name,
-                    "status": "PASS" if r.passed else "FAIL",
-                    "duration": r.duration,
-                    "details": r.details,
-                }
-                for r in report.results
-            ],
-        }
-
-        return test_results
-
-    except Exception as e:
-        logger.error(f"❌ Comprehensive integration suite failed: {e}")
-        return {
-            "tests_run": 0,
-            "tests_passed": 0,
-            "tests_failed": 1,
-            "tests_error": 0,
-            "total_time": 0,
-            "integration_status": {"error": str(e)},
-            "performance_metrics": {},
-            "test_details": [
-                {
-                    "test": "Comprehensive Integration Suite",
-                    "status": "ERROR",
-                    "duration": 0,
-                    "details": str(e),
-                }
-            ],
-        }
-
-
-def run_basic_tests_only(logger: logging.Logger) -> dict:
-    """Run basic tests when integration components are not available"""
-    logger.info("🔧 Running Basic Tests Only")
-    logger.info("-" * 40)
-
-    try:
-        # Initialize basic auth service
-        auth_service = AuthService()
-
-        # Run basic functionality tests
-        basic_results = run_basic_functionality_tests(auth_service, logger)
-
-        # Run performance tests
-        perf_results = run_performance_tests(auth_service, logger)
-
-        # Run integration tests
-        integration_results = run_integration_tests(auth_service, logger)
-
-        # Combine results
-        total_tests = (
-            basic_results["tests_run"]
-            + perf_results["tests_run"]
-            + integration_results["tests_run"]
-        )
-        total_passed = (
-            basic_results["tests_passed"]
-            + perf_results["tests_passed"]
-            + integration_results["tests_passed"]
-        )
-        total_failed = (
-            basic_results["tests_failed"]
-            + perf_results["tests_failed"]
-            + integration_results["tests_failed"]
-        )
-
-        combined_results = {
-            "tests_run": total_tests,
-            "tests_passed": total_passed,
-            "tests_failed": total_failed,
-            "tests_error": 0,
-            "total_time": 0,
-            "integration_status": {"basic_mode": True, "integration_components": False},
-            "performance_metrics": {},
-            "test_details": (
-                basic_results["test_details"]
-                + perf_results["test_details"]
-                + integration_results["test_details"]
-            ),
-        }
-
-        # Cleanup
-        auth_service.shutdown()
-
-        return combined_results
-
-    except Exception as e:
-        logger.error(f"❌ Basic tests failed: {e}")
-        return {
-            "tests_run": 0,
-            "tests_passed": 0,
-            "tests_failed": 1,
-            "tests_error": 0,
-            "total_time": 0,
-            "integration_status": {"error": str(e)},
-            "performance_metrics": {},
-            "test_details": [
-                {
-                    "test": "Basic Tests",
-                    "status": "ERROR",
-                    "duration": 0,
-                    "details": str(e),
-                }
-            ],
-        }
-
-
-def generate_test_report(test_results: dict, logger: logging.Logger) -> dict:
-    """Generate comprehensive test report"""
-    logger.info("📊 Generating Test Report")
-    logger.info("-" * 40)
-
-    # Calculate success rate
-    total_tests = test_results["tests_run"]
-    passed_tests = test_results["tests_passed"]
-    failed_tests = test_results["tests_failed"]
-    error_tests = test_results.get("tests_error", 0)
-
-    success_rate = (passed_tests / total_tests * 100) if total_tests > 0 else 0
-
-    # Generate report
-    report = {
-        "report_id": f"auth_integration_report_{int(time.time())}",
-        "timestamp": datetime.now().isoformat(),
-        "test_summary": {
-            "total_tests": total_tests,
-            "passed_tests": passed_tests,
-            "failed_tests": failed_tests,
-            "error_tests": error_tests,
-            "success_rate": f"{success_rate:.1f}%",
-        },
-        "test_results": test_results,
-        "recommendations": [],
-        "status": "PASS" if failed_tests == 0 and error_tests == 0 else "FAIL",
-    }
-
-    # Generate recommendations
-    if failed_tests > 0:
-        report["recommendations"].append(f"Investigate {failed_tests} failed tests")
-
-    if error_tests > 0:
-        report["recommendations"].append(f"Fix {error_tests} test errors")
-
-    if success_rate < 90:
-        report["recommendations"].append(
-            "Overall success rate below 90% - review system"
-        )
-
-    if success_rate >= 95:
-        report["recommendations"].append(
-            "Excellent test results - system ready for production"
-        )
-
-    return report
-
-
-def save_test_report(report: dict, logger: logging.Logger):
-    """Save test report to file"""
-    try:
-        timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-        filename = f"auth_integration_report_{timestamp}.json"
-
-        with open(filename, "w") as f:
-            json.dump(report, f, indent=2, default=str)
-
-        logger.info(f"✅ Test report saved to {filename}")
-
-    except Exception as e:
-        logger.error(f"❌ Failed to save test report: {e}")
+sys.path.append(str(Path(__file__).resolve().parents[3]))
+
+from auth_integration_test_setup import setup_logging
+from auth_integration_test_execution import run_comprehensive_integration_suite
+from auth_integration_test_reporting import generate_test_report, save_test_report
 
 
 def main():
-    """Main integration test runner"""
+    """Main integration test runner."""
     logger = setup_logging()
 
     logger.info("=" * 80)
     logger.info("🚀 V2 AUTHENTICATION INTEGRATION TEST SUITE")
     logger.info("=" * 80)
     logger.info(f"Started at: {datetime.now()}")
-    logger.info(f"Agent-2: AI & ML Integration Specialist")
-    logger.info(f"Task: Begin integration tests for services_v2/auth. Report in 60m.")
+    logger.info("Agent-2: AI & ML Integration Specialist")
+    logger.info("Task: Begin integration tests for services_v2/auth. Report in 60m.")
     logger.info("=" * 80)
 
-    try:
-        # Run comprehensive integration tests
-        start_time = time.time()
-        test_results = run_comprehensive_integration_suite(logger)
-        total_time = time.time() - start_time
+    start_time = time.time()
+    test_results = run_comprehensive_integration_suite(logger)
+    total_time = time.time() - start_time
 
-        # Log results summary
-        logger.info("\n" + "=" * 80)
-        logger.info("🏁 INTEGRATION TEST SUITE COMPLETED")
-        logger.info("=" * 80)
-        logger.info(f"Total Tests: {test_results['tests_run']}")
-        logger.info(f"Passed: {test_results['tests_passed']}")
-        logger.info(f"Failed: {test_results['tests_failed']}")
-        logger.info(f"Errors: {test_results.get('tests_error', 0)}")
-        logger.info(f"Total Time: {total_time:.2f}s")
+    logger.info("\n" + "=" * 80)
+    logger.info("🏁 INTEGRATION TEST SUITE COMPLETED")
+    logger.info("=" * 80)
+    logger.info(f"Total Tests: {test_results['tests_run']}")
+    logger.info(f"Passed: {test_results['tests_passed']}")
+    logger.info(f"Failed: {test_results['tests_failed']}")
+    logger.info(f"Errors: {test_results.get('tests_error', 0)}")
+    logger.info(f"Total Time: {total_time:.2f}s")
 
-        if test_results["tests_run"] > 0:
-            success_rate = (
-                test_results["tests_passed"] / test_results["tests_run"]
-            ) * 100
-            logger.info(f"Success Rate: {success_rate:.1f}%")
+    if test_results["tests_run"] > 0:
+        success_rate = (
+            test_results["tests_passed"] / test_results["tests_run"]
+        ) * 100
+        logger.info(f"Success Rate: {success_rate:.1f}%")
 
-        # Generate and save report
-        report = generate_test_report(test_results, logger)
-        save_test_report(report, logger)
+    report = generate_test_report(test_results, logger)
+    save_test_report(report, logger)
 
-        # Final status
-        if report["status"] == "PASS":
-            logger.info("🎉 ALL TESTS PASSED - System ready for production!")
-        else:
-            logger.warning("⚠️ Some tests failed - Review recommendations")
+    if report["status"] == "PASS":
+        logger.info("🎉 ALL TESTS PASSED - System ready for production!")
+    else:
+        logger.warning("⚠️ Some tests failed - Review recommendations")
 
-        logger.info("=" * 80)
-
-        return report
-
-    except Exception as e:
-        logger.error(f"❌ Integration test suite failed: {e}")
-
-        # Generate error report
-        error_results = {
-            "tests_run": 0,
-            "tests_passed": 0,
-            "tests_failed": 0,
-            "tests_error": 1,
-            "total_time": 0,
-            "integration_status": {"error": str(e)},
-            "performance_metrics": {},
-            "test_details": [
-                {
-                    "test": "Integration Test Suite",
-                    "status": "ERROR",
-                    "duration": 0,
-                    "details": str(e),
-                }
-            ],
-        }
-
-        report = generate_test_report(error_results, logger)
-        save_test_report(report, logger)
-
-        return report
+    logger.info("=" * 80)
+    return report
 
 
 if __name__ == "__main__":
